+++ src/dune/dune_file.mli
@@ -152,9 +152,6 @@ module Mode_conf : sig
 
     val decode : t Dune_lang.Decoder.t
 
-    (** Byte inherited, Best is requested *)
-    val default : Loc.t -> t
-
     module Details : sig
       type t = Kind.t option
     end